Update RetrofitInterfaces : Migrated Interfaces
